'schema' and 'host' fields are now validated, reviewed Swagger model, code refactoring, add jUnit testing
