+++ .idea/libraries/Gradle__net_portswigger_burp_extender_burp_extender_api_1_7_22.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="Gradle: net.portswigger.burp.extender:burp-extender-api:1.7.22">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/net.portswigger.burp.extender/burp-extender-api/1.7.22/79325774249c57f97d7adb9f0101ca3338c865a5/burp-extender-api-1.7.22.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/net.portswigger.burp.extender/burp-extender-api/1.7.22/8c3b2b46d7484a5de0099efe70f5883430146dcc/burp-extender-api-1.7.22-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
