+++ build.gradle
@@ -5,18 +5,17 @@ repositories {
 }
 
 dependencies {
+    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.9'
+    compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.8.9' 
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.0'
     compile group: 'net.portswigger.burp.extender', name: 'burp-extender-api', version: '1.7.22'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.9'
-    compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.8.9'    
+
+    testCompile 'junit:junit:4.12'
+    testCompile group: 'org.slf4j', name: 'slf4j-simple', version: '1.6.1'
 }
 
-sourceSets {
-    main {
-        java {
-            srcDir 'src'
-        }
-    }
+test {
+    testLogging.showStandardStreams = true   
 }
 
 task fatJar(type: Jar) {
