+++ src/main/java/swurg/model/Info.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright (C) 2016 Alexandre Teyar
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package swurg.model;
+
+public class Info {
+    private String version;
+    private String title;
+
+    public Info(String version, String title) {
+        this.version = version;
+        this.title = title;
+    }
+
+    public String getVersion() {
+        return this.version;
+    }
+
+    public String getTitle() {
+        return this.title;
+    }
+}
