+++ src/main/java/swurg/process/Loader.java
@@ -1,0 +1,43 @@
+/**
+ * Copyright (C) 2016 Alexandre Teyar
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package swurg.process;
+
+import com.google.gson.Gson;
+import swurg.model.REST;
+
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+
+public class Loader {
+    public REST process(File file) throws FileNotFoundException {
+        if (file == null) {
+            throw new IllegalArgumentException("No file specified");
+        }
+
+        if (!file.exists()) {
+            throw new IllegalArgumentException("File doesn't exist!");
+        }
+
+        Gson gson = new Gson();
+
+        BufferedReader bufferedReader = new BufferedReader(new FileReader(file));
+
+        return gson.fromJson(bufferedReader, REST.class);
+    }
+}
