+++ src/test/resources/log4j.properties
@@ -1,0 +1,5 @@
+log4j.rootLogger=DEBUG, STDOUT
+log4j.logger.deng=INFO
+log4j.appender.STDOUT=org.apache.log4j.ConsoleAppender
+log4j.appender.STDOUT.layout=org.apache.log4j.PatternLayout
+log4j.appender.STDOUT.layout.ConversionPattern=%5p [%t] (%F:%L) - %m%n
