+++ src/test/resources/testApi.json
@@ -1,0 +1,158 @@
+{
+  "swagger": "2.0",
+  "info": {
+    "description": "Test API for swurg.\n",
+    "version": "1.0.0",
+    "title": "Test API..."
+  },
+  "host": "localhost",
+  "basePath": "/api/test/v1",
+  "schemes": [
+    "https"
+  ],
+  "paths": {
+    "/stuff": {
+      "get": {
+        "summary": "Get current basket",
+        "description": "Returns the current basket contents.",
+        "produces": [
+          "application/json"
+        ],
+        "responses": {
+          "200": {
+            "description": "Successful Operation",
+            "schema": {
+              "type": "object",
+              "properties": {
+                "pricing": {
+                  "$ref": "#/definitions/Pricing"
+                },
+                "lines": {
+                  "type": "array",
+                  "items": {
+                    "$ref": "#/definitions/LineItem"
+                  }
+                },
+                "payment_reference": {
+                  "type": "string"
+                },
+                "payment_gateway_url": {
+                  "type": "string"
+                }
+              }
+            }
+          },
+          "404": {
+            "description": "Basket does not exist"
+          }
+        }
+      }
+    },
+    "/stuff/deliveryoption": {
+      "get": {
+        "description": "get order delivery options.",
+        "produces": [
+          "application/json"
+        ],
+        "responses": {
+          "200": {
+            "description": "successful operation",
+            "schema": {
+              "type": "array",
+              "items": {
+                "$ref": "#/definitions/DeliveryOption"
+              }
+            }
+          },
+          "404": {
+            "description": "Not Found"
+          }
+        }
+      }
+    },
+    "/stuff/deliveryaddress": {
+      "get": {
+        "description": "Get basket Shipping Address Options.",
+        "produces": [
+          "application/json"
+        ],
+        "responses": {
+          "200": {
+            "description": "successful operation",
+            "schema": {
+              "type": "array",
+              "items": {
+                "$ref": "#/definitions/Address"
+              }
+            }
+          },
+          "404": {
+            "description": "Not Found"
+          }
+        }
+      }
+    }
+  },
+  "definitions": {
+    "DeliveryOption": {
+      "type": "object",
+      "properties": {
+        "id": {
+          "type": "integer"
+        },
+        "title": {
+          "type": "string"
+        },
+        "description": {
+          "type": "string"
+        },
+        "price": {
+          "type": "number",
+          "format": "float"
+        },
+        "currency_code": {
+          "type": "string"
+        }
+      }
+    },
+    "Address": {
+      "type": "object",
+      "properties": {
+        "id": {
+          "type": "integer",
+          "format": "int32"
+        },
+        "title": {
+          "type": "string"
+        },
+        "firstname": {
+          "type": "string"
+        },
+        "lastname": {
+          "type": "string"
+        },
+        "thoroughfare": {
+          "type": "string"
+        },
+        "premise": {
+          "type": "string"
+        },
+        "dependent_locality": {
+          "type": "string"
+        },
+        "locality": {
+          "type": "string"
+        },
+        "administrative_area": {
+          "type": "string"
+        },
+        "postal_code": {
+          "type": "string"
+        },
+        "country": {
+          "type": "string"
+        }
+      }
+    }
+  }
+}
